NA_Rename_api_operations_to_api_operation
