+++ lib/uiza.rb
@@ -15,13 +15,13 @@ require "uiza/error/service_unavailable_error"
 require "uiza/error/client_error"
 require "uiza/error/server_error"
 
-require "uiza/api_operations/create"
-require "uiza/api_operations/add"
-require "uiza/api_operations/retrieve"
-require "uiza/api_operations/list"
-require "uiza/api_operations/update"
-require "uiza/api_operations/delete"
-require "uiza/api_operations/remove"
+require "uiza/api_operation/create"
+require "uiza/api_operation/add"
+require "uiza/api_operation/retrieve"
+require "uiza/api_operation/list"
+require "uiza/api_operation/update"
+require "uiza/api_operation/delete"
+require "uiza/api_operation/remove"
 
 require "uiza/api_resource/entity_resource"
