+++ lib/uiza/api_operation/delete.rb
@@ -1,13 +1,13 @@
 module Uiza
-  module APIOperations
-    module Remove
-      def remove id
+  module APIOperation
+    module Delete
+      def delete id
         url = "https://#{Uiza.workspace_api_domain}/api/public/v3/#{self::OBJECT_API_PATH}"
         method = :delete
         headers = {"Authorization" => Uiza.authorization}
         params = {id: id}
 
-        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:remove]
+        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:delete]
         uiza_client.execute_request
       end
     end
