+++ lib/uiza/api_operation/list.rb
@@ -1,13 +1,12 @@
 module Uiza
-  module APIOperations
-    module Retrieve
-      def retrieve id
+  module APIOperation
+    module List
+      def list params = {}
         url = "https://#{Uiza.workspace_api_domain}/api/public/v3/#{self::OBJECT_API_PATH}"
         method = :get
         headers = {"Authorization" => Uiza.authorization}
-        params = {id: id}
 
-        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:retrieve]
+        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:list]
         uiza_client.execute_request
       end
     end
