+++ lib/uiza/api_operation/remove.rb
@@ -1,12 +1,13 @@
 module Uiza
-  module APIOperations
-    module Add
-      def add params
+  module APIOperation
+    module Remove
+      def remove id
         url = "https://#{Uiza.workspace_api_domain}/api/public/v3/#{self::OBJECT_API_PATH}"
-        method = :post
+        method = :delete
         headers = {"Authorization" => Uiza.authorization}
+        params = {id: id}
 
-        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:add]
+        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:remove]
         uiza_client.execute_request
       end
     end
