+++ lib/uiza/api_operation/update.rb
@@ -1,12 +1,12 @@
 module Uiza
-  module APIOperations
-    module Create
-      def create params
+  module APIOperation
+    module Update
+      def update params
         url = "https://#{Uiza.workspace_api_domain}/api/public/v3/#{self::OBJECT_API_PATH}"
-        method = :post
+        method = :put
         headers = {"Authorization" => Uiza.authorization}
 
-        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:create]
+        uiza_client = UizaClient.new url, method, headers, params, self::OBJECT_API_DESCRIPTION_LINK[:update]
         response = uiza_client.execute_request
 
         self.retrieve response.id
