+++ lib/uiza/category.rb
@@ -1,7 +1,7 @@
 module Uiza
   class Category
-    extend Uiza::APIOperations::Create
-    extend Uiza::APIOperations::Retrieve
+    extend Uiza::APIOperation::Create
+    extend Uiza::APIOperation::Retrieve
 
     OBJECT_API_PATH = "media/metadata".freeze
   end
