+++ lib/uiza/storage.rb
@@ -1,9 +1,9 @@
 module Uiza
   class Storage
-    extend Uiza::APIOperations::Add
-    extend Uiza::APIOperations::Retrieve
-    extend Uiza::APIOperations::Update
-    extend Uiza::APIOperations::Remove
+    extend Uiza::APIOperation::Add
+    extend Uiza::APIOperation::Retrieve
+    extend Uiza::APIOperation::Update
+    extend Uiza::APIOperation::Remove
 
     OBJECT_API_PATH = "media/storage".freeze
   end
